Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Authlib
Flask-Mail
python-dotenv
Flask-CORS
requests 